Added support for the VS 2017 header (#60)
Due to a slight variation in the Visual Studio 2017 file header, which I missed, we couldn't properly parse the version information from a 2017 file header. The portion containing the line for the major version used to indicate the solution icon does not contain the word "Version" like in 2019 and 2022. This change allows us to handle the 2017 variation.