Allow latest SDK again and patch on macOS (#160)
This change reverts a portion of my last change to allow us to use the
latest .NET 7 SDK again. The issue is on macOS where `dotnet format` on
7.0.101 is broken, so i've included a script that patches global.json on
macOS to disable the `rollForward` option. I've also installed 2
versions of the .NET 7 SDK on build agents to allow both variations to
work.