Upgrade to .NET 8 (#214)
This change upgrades the project to use the .NET 8 SDK and adds the
`net8.0` target to all projects. I've also removed support for .NET 7
and updated the version to 2.0. It also addresses the new analyzer
warnings introduced as well as conforms to new C# 12 feature
recommendations imposed by the new analyzers. Using the new artifacts
support for output.