Migrate to slnx (#344)
This change migrates the project to use the new slnx format. See
[here](https://devblogs.microsoft.com/dotnet/introducing-slnx-support-dotnet-cli/)
and
[here](https://devblogs.microsoft.com/visualstudio/new-simpler-solution-file-format/)
for more details.

Essentially, I just:

```powershell
dotnet sln migrate
rm ./slnup.sln
```

As there is no header information in this file anymore (that i'm aware
of), this may very well render the SlnUp project unnecessary, but it
also begs the question as to how the Visual Studio Version Selector will
be able to do its job. To be continued...